Create genshinRcode for Milestone 5
I analyzed two quantitative variables from my dataset: base HP and base Attack. The mean and median for both variables were very close to each other, indicating that the data is symmetrically distributed without significant outliers. The HP values had a mean of 950.42 and a median of 951, while the Attack values had a mean of 19.52 and a median of 19. It seems like the distributions are fairly balanced. The standard deviation for HP which was 116.46, was higher showing greater variation in character HP values. For the Attack standard deviation (3.59) was small, demonstrating that the attack values are closely grouped around the average. Overall, both variables show consistent patterns with minimal skew, so a trimmed mean was not necessary for this. :)