Milestone#6 with Rcode
Correlation coefficient: 0.176

There’s a very weak positive linear relationship, and because of that, characters with higher HP tend to have slightly higher ATK, but it’s not a strong correlation it sadly.


The scatterplot of HP vs ATK for Genshin characters demonstrates a weak positive linear relationship. With a correlation coefficient of 0.176, suggesting that as HP increases, ATK may increase slightly as well but the association is still weak. The data seems to be scattered very widely.